Refactored pet summoning from DemoPets
