+++ src/parser/warlock/demonology/modules/pets/DemoPets/index.js
@@ -4,14 +4,10 @@ import SPELLS from 'common/SPELLS';
 import RESOURCE_TYPES from 'game/RESOURCE_TYPES';
 
 import Timeline from '../Timeline';
-import { TimelinePet, DESPAWN_REASONS, META_CLASSES, META_TOOLTIPS } from '../TimelinePet';
+import { DESPAWN_REASONS, META_CLASSES, META_TOOLTIPS } from '../TimelinePet';
 import PetDamage from '../PetDamage';
-import { isPermanentPet } from '../helpers';
 import PETS from '../PETS';
-import { SUMMON_TO_SPELL_MAP } from '../CONSTANTS';
 
-const BUFFER = 150;
-const DEMONIC_POWER_DURATION = 15000;
 const debug = false;
 const test = false;
 
@@ -53,14 +49,8 @@ class DemoPets extends Analyzer {
   timeline = new Timeline();
 
   _hasDemonicConsumption = false;
-  _lastIDtick = null;
-  _lastSpendResource = null;
   _lastImplosionCast = null;
   _implosionTargetsHit = [];
-  _lastPlayerPosition = {
-    x: 0,
-    y: 0,
-  };
   _lastDemonicTyrantCast = null;
   _wildImpIds = []; // important for different handling of duration, these IDs change from log to log
   _petsAffectedByDemonicTyrant = []; // dynamic because of talents
@@ -74,41 +64,7 @@ class DemoPets extends Analyzer {
 
   // Pet timeline
 
-  on_byPlayer_summon(event) {
-    const petInfo = this._getPetInfo(event.targetID);
-    if (!petInfo) {
-      debug && this.error('Summoned unknown pet', event);
-      return;
-    }
-    if (isPermanentPet(petInfo.guid)) {
-      test && this.log('Permanent pet summon');
-      this.timeline.tryDespawnLastPermanentPet(event.timestamp);
-    }
-    const pet = new TimelinePet(petInfo,
-                                event.targetID,
-                                event.targetInstance,
-                                event.timestamp,
-                                this._getPetDuration(event.targetID),
-                                this._getSummonSpell(event),
-                                event.ability.guid);
-    if (this._wildImpIds.includes(pet.id)) {
-      // Wild Imps need few additional properties
-      pet.setWildImpProperties(this._lastPlayerPosition);
-    }
-    test && this.log('Pet summoned', pet);
-    this.timeline.addPet(pet);
-    pet.pushHistory(event.timestamp, 'Summoned', event);
-    if (pet.summonedBy === SPELLS.INNER_DEMONS_TALENT.id) {
-      this._lastIDtick = event.timestamp;
-    }
-  }
-
-  on_byPlayer_spendresource(event) {
-    this._lastSpendResource = event.timestamp;
-  }
-
   on_byPlayer_cast(event) {
-    this._updatePlayerPosition(event);
     if (event.ability.guid === SPELLS.IMPLOSION_CAST.id) {
       this._handleImplosionCast(event);
     }
@@ -226,26 +182,6 @@ class DemoPets extends Analyzer {
         imp.expectedDespawn = imp.spawn + PETS.WILD_IMP_HOG.duration + actualBuffTime;
         imp.pushHistory(event.timestamp, 'Updated expected despawn from', old, 'to', imp.expectedDespawn);
       });
-  }
-
-  // to update player position more accurately (based on DistanceMoved)
-  // important, since Wild Imp summons uses player position as default (not entirely accurate, as they're spawned around player, not exactly on top of it, but that's as close as I'm gonna get)
-  // needed for Implosion - there's a possibility that Wild Imps don't cast anything between their 'summon' and Implosion, therefore I wouldn't get their position
-
-  on_toPlayer_damage(event) {
-    this._updatePlayerPosition(event);
-  }
-
-  on_toPlayer_energize(event) {
-    this._updatePlayerPosition(event);
-  }
-
-  on_toPlayer_heal(event) {
-    this._updatePlayerPosition(event);
-  }
-
-  on_toPlayer_absorbed(event) {
-    this._updatePlayerPosition(event);
   }
 
   // API
@@ -366,35 +302,6 @@ class DemoPets extends Analyzer {
     return pet;
   }
 
-  _getPetGuid(id, isGuid = false) {
-    const pet = this._getPetInfo(id, isGuid);
-    return pet ? pet.guid : null;
-  }
-
-  _getPetDuration(id, isGuid = false) {
-    const pet = this._getPetInfo(id, isGuid);
-    if (!pet) {
-      debug && this.error(`NewPets._getPetDuration() called with nonexistant pet ${isGuid ? 'gu' : ''}id ${id}`);
-      return -1;
-    }
-    if (isPermanentPet(pet.guid)) {
-      debug && this.log('Called _getPetDuration() for permanent pet guid', pet);
-      return Infinity;
-    }
-    if (!PETS[pet.guid]) {
-      debug && this.error('Encountered pet unknown to PET_INFO.js', pet);
-      return -1;
-    }
-    // for imps, take Demonic Tyrant in consideration
-    // if player doesn't have the buff, it's 15 seconds
-    if (this._wildImpIds.includes(pet.id) && this.selectedCombatant.hasBuff(SPELLS.DEMONIC_POWER.id)) {
-      // if player has the buff, it takes the remaining buff time + 15 seconds
-      const remainingBuffTime = (this._lastDemonicTyrantCast + DEMONIC_POWER_DURATION) - this.owner.currentTimestamp;
-      return PETS[pet.guid].duration + remainingBuffTime;
-    }
-    return PETS[pet.guid].duration;
-  }
-
   _initializeWildImps() {
     // there's very little possibility these statements wouldn't return an object, Hand of Guldan is a key part of rotation
     this._wildImpIds.push(this._toId(PETS.WILD_IMP_HOG.guid));
@@ -431,30 +338,8 @@ class DemoPets extends Analyzer {
     ];
   }
 
-  _getSummonSpell(event) {
-    if (!SUMMON_TO_SPELL_MAP[event.ability.guid]) {
-      if (event.timestamp <= this._lastIDtick + BUFFER) {
-        return SPELLS.INNER_DEMONS_TALENT.id;
-      }
-      if (this.selectedCombatant.hasBuff(SPELLS.NETHER_PORTAL_BUFF.id) && event.timestamp <= this._lastSpendResource + BUFFER) {
-        return SPELLS.NETHER_PORTAL_TALENT.id;
-      }
-      debug && this.error('Unknown source of summon event', event);
-      return -1;
-    }
-    return SUMMON_TO_SPELL_MAP[event.ability.guid];
-  }
-
   _getDistance(x1, y1, x2, y2) {
     return Math.sqrt((x1 - x2) * (x1 - x2) + (y1 - y2) * (y1 - y2));
-  }
-
-  _updatePlayerPosition(event) {
-    if (!event.x || !event.y) {
-      return;
-    }
-    this._lastPlayerPosition.x = event.x;
-    this._lastPlayerPosition.y = event.y;
   }
 }
